Fix sql exists bug
Quote.exists() was only checking if date in row was unique. It did take
into account if date and symbol name where unique. So when you wanted to
add quotes froma different symbol, say MSFT, then we would run
Quote.exists(date). It would already see that there was a row with said
date and not wirte another row. But that previous row was for a
different symbol.